requirement.txt
fastapi[standard]
pydantic
sqlalchemy
psycopg[binary]
alembic
python-dotenv
email-validator
